Add files via upload
The Credit Card Fraud Detection Machine Learning project is a solution designed to detect fraudulent credit card transactions. The project uses various algorithms and techniques, such as anomaly detection, to analyze transaction data and identify patterns that may indicate fraud. The project may include pre-processing of credit card transaction data, feature engineering, model selection and evaluation, and deployment to a production environment. The goal of the project is to help financial institutions reduce the risk of credit card fraud and improve their overall fraud detection capabilities. The source code, documentation, and other resources related to the project are hosted on a code repository such as GitHub, allowing for collaboration and community contributions.